Set up add subcategory functionality
