+++ src/Components/AddSubcategoryForm/AddSubcategoryForm.js
@@ -1,57 +1,15 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
+import UseAddSubcategoryForm from './UseAddSubcategoryForm'
 
 
-function AddSubcategoryForm() {
+function AddSubcategoryForm(props) {
+    const [subcatTitleInput, setSubcatTitleInput] = useState('')
+    const [subcatImageInput, setSubcatImageInput] = useState('')
 
-    const imageSelection = [
-        { title: "hello", image: "👋" },
-        { title: "bank", image: "🏦" },
-        { title: "morning", image: "🌅" },
-        { title: "afternoon", image: "🏙" },
-        { title: "night", image: "🌃" },
-        { title: "yes", image: "👍" },
-        { title: "no", image: "🚫" },
-        { title: "great", image: "😀" },
-        { title: "perfect", image: "👌" },
-        { title: "card", image: "💳" },
-        { title: "thanks", image: "🙏" },
-        { title: "withdrawal", image: "💰" },
-        { title: "deposit", image: "💸" },
-        { title: "memo", image: "📝" },
-        { title: "bank-note", image: "💵" },
-        { title: "great-day", image: "😄" },
-        { title: "x-hands", image: "🙅‍♀️" },
-        { title: "party", image: "🎉" },
-        { title: "yes", image: "👍" },
-        { title: "no", image: "🚫" },
-        { title: "sup", image: "👋" },
-        { title: "yo", image: "🪀" },
-        { title: "hundred", image: "💯" },
-        { title: "rabbit", image: "🐇" },
-        { title: "joke-1", image: "🤣" },
-        { title: "joke-2", image: "🙉" },
-        { title: "dog", image: "🐶" },
-        { title: "rtj", image: "🎹" },
-        { title: "school", image: "🎒" },
-        { title: "drink", image: "🥃" },
-        { title: "peace", image: "✌️" },
-        { title: "wave", image: "👋" },
-        { title: "felicia", image: "👩" },
-        { title: "sos", image: "🆘" },
-        { title: "Greetings", image: "🤝" },
-        { title: "Common", image: "😎" },
-        { title: "Transactions", image: "💰" },
-        { title: "Goodbyes", image: "👋" },
-        { title: "Jokes", image: "😜" },
-        { title: "Joe", image: "😊" },
-    ]
+    const subCategoryImages = UseAddSubcategoryForm()
 
-    const optionItems = imageSelection.map((selection) =>
-        <option value={selection.title}>{selection.image + selection.title}
-        </option>
-    );
-
+    
     return (
         <>
             <form 
@@ -68,40 +26,39 @@ function AddSubcategoryForm() {
                         name='text-input'
                         className='phrase-input'
                         onChange={(e) => {
-                        // setPhraseInput(e.target.value)
+                        setSubcatTitleInput(e.target.value)
                     }}
-                    // value={phraseInput}
+                    value={subcatTitleInput}
                 ></input>
                 <select 
                     name="icon" 
                     className="phrase-input" 
                     id="icon-select" 
-                    // onChange={(e) => { setImageInput(e.target.value) }}
+                    onChange={(e) => { setSubcatImageInput(e.target.value) }}
                 >
                     <option value="">--Please Choose an Icon--</option>
-                    {optionItems}
+                    {subCategoryImages}
                 </select>
                 <section className='play-save-container'>
-                <Link 
-                    to="/" 
-                    style={{ textDecoration: 'none' }}
-                    onClick={e => {
-                        // handleFormSubmit(e)
-                        // setPhraseInput('')
-                        // setImageInput()
-                    }
-                    }
-                >
-                    <button
-                      className="save-and-play-buttons"
-                      value='save'
-                      name='save-button'
-                      label={'Save'}
-                  >
-                    Save
-                  </button>
-              </Link>
-          </section>
+                    <Link 
+                        to="/" 
+                        style={{ textDecoration: 'none' }}
+                        onClick={e => {
+                            // handleFormSubmit(e)
+                            setSubcatTitleInput('')
+                            setSubcatImageInput()
+                            }
+                        }
+                    >
+                        <button
+                            className="save-and-play-buttons"
+                            value='save'
+                            name='save-button'
+                            label={'Save'}
+                        >Save
+                        </button>
+                    </Link>
+                </section>
             </form>
         </>
   )
